* Moved NameGenerator to Utils, so it can be acessed from different packages
