+++ lib/mosaic-utils/src/main/java/org/eclipse/mosaic/lib/util/NameGenerator.java
@@ -13,7 +13,7 @@
  * Contact: mosaic@fokus.fraunhofer.de
  */
 
-package org.eclipse.mosaic.fed.mapping.ambassador;
+package org.eclipse.mosaic.lib.util;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -61,7 +61,10 @@ public class NameGenerator {
         }
     }
 
-    static void reset() {
+    /**
+     * Resets all counters for the name generators.
+     */
+    public static void reset() {
         vehicleCounter.set(0);
         rsuCounter.set(0);
         tmcCounter.set(0);
